Fix bug in handling of canonical aliases
continues 00d18de46e2f4bf21c0ab4b2276b8942eb4fa32c

+++ src/lib/Reconciler.ts
@@ -337,7 +337,12 @@ export default class Reconciler {
 
     const existingByTag = expected.tag && this.resolveTag(expected.tag);
     const existingByAlias = await this.resolveAlias(alias);
-    if (existingByTag && existingByAlias && existingByAlias !== existingByTag) {
+    if (
+      existingByTag &&
+      existingByAlias &&
+      existingByAlias !== existingByTag &&
+      alias.endsWith(this.plan.homeserver)
+    ) {
       info("🏷️ Delete alias", { alias });
       await this.matrix.deleteRoomAlias(alias);
     }
